Move config under the 'home' folder
In an attempt to generalize my dot config install scripts I am creating
a 'home' folder, moving the config files into this folder maintaining
the folder structure that should be in the users' home directory. I then
can simple loop through all the items in the 'home' folder and link them
to the file/folder that exists in the current projects home folder.